% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addACSCalculatedVariables.R
\name{addACSCalculatedVariables}
\alias{addACSCalculatedVariables}
\title{addACSCalculatedVariables}
\usage{
addACSCalculatedVariables(base_data)
}
\arguments{
\item{base_data}{dataframe with ADA-PARC base variables}
}
\value{
a DF with ADA-PARC base and calculated variables
}
\description{
Calculate variables and add to dataset, based on ACS variables
}
