% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCalculatedVariablesForBaseDemographics.R
\name{addCalculatedVariablesForBaseDemographics}
\alias{addCalculatedVariablesForBaseDemographics}
\title{addCalculatedVariablesForBaseDemographics}
\usage{
addCalculatedVariablesForBaseDemographics(base_data)
}
\arguments{
\item{base_data}{dataframe with ADA-PARC base variables}
}
\value{
a DF with ADA-PARC base and calculated variables
}
\description{
Add Calculated Variables To Base Data for Demographics
}
