% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCalculatedVariablesForBaseEconomic.R
\name{addCalculatedVariablesForBaseEconomic}
\alias{addCalculatedVariablesForBaseEconomic}
\title{addCalculatedVariablesForBaseEconomic}
\usage{
addCalculatedVariablesForBaseEconomic(base_data)
}
\arguments{
\item{base_data}{dataframe with ADA-PARC base variables}
}
\value{
a DF with ADA-PARC base and calculated variables
}
\description{
Add Calculated Variables To Base Data for Economic
}
