% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCalculatedVariablesForBaseParticipation.R
\name{addCalculatedVariablesForBaseParticipation}
\alias{addCalculatedVariablesForBaseParticipation}
\title{addCalculatedVariablesForBaseParticipation}
\usage{
addCalculatedVariablesForBaseParticipation(base_data)
}
\arguments{
\item{base_data}{dataframe with ADA-PARC base variables}
}
\value{
a DF with ADA-PARC base and calculated variables
}
\description{
Add Calculated Variables To Base Data for Participation
}
