% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCalculatedVariablesToBaseData.R
\name{addCalculatedVariablesToBaseData}
\alias{addCalculatedVariablesToBaseData}
\title{addCalculatedVariablesToBaseData}
\usage{
addCalculatedVariablesToBaseData(ada_parc_base_and_calc_vars_map, base_data)
}
\arguments{
\item{ada_parc_base_and_calc_vars_map}{df mapping ADA_PARC base and calculated variables}

\item{base_data}{dataframe with ADA-PARC base variables}
}
\value{
a tibble with ADA-PARC base and calculated variables
}
\description{
Add Calculated Variables To Base Data
}
