% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadACSPlace.R
\name{downloadACSPlace}
\alias{downloadACSPlace}
\title{downloadACSPlace}
\usage{
downloadACSPlace(tables, geography = "place", year, survey = "acs5")
}
\arguments{
\item{tables}{array of strings, corresponding to the US Census tables to be downloaded}

\item{geography}{the geography of the request; can be either "state" or "us".}

\item{year}{year of data to download}

\item{survey}{specification of ACS survey type (ex. "acs5" or "acs1")}
}
\value{
A dataframe of downloaded ACS data, in wide format.
}
\description{
Downloads and Formats ACS Data for the place geography
}
