% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadACSPlacePopulationEstimate.R
\name{downloadACSPlacePopulationEstimate}
\alias{downloadACSPlacePopulationEstimate}
\title{downloadACSPlacePopulationEstimate}
\usage{
downloadACSPlacePopulationEstimate(year)
}
\arguments{
\item{year}{the year for the lookup}
}
\value{
a dataframe with GEOID, Name, and Population of places in the US
}
\description{
Download the ACS Population Estimates for place
}
\details{
Downloads and filters ACS population estimates at the "place" level across the US
}
