% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadACSTracts.R
\name{downloadACSTracts}
\alias{downloadACSTracts}
\title{downloadStateWideAcs}
\usage{
downloadACSTracts(
  tables,
  geography = "tract",
  year,
  survey = "acs5",
  fips_codes_for_lookup
)
}
\arguments{
\item{tables}{array of strings, corresponding to the US Census tables to be downloaded}

\item{geography}{the geography of the request; can be either "state" or "us".}

\item{year}{year of data to download}

\item{survey}{specification of ACS survey type (ex. "acs5" or "acs1")}

\item{fips_codes_for_lookup}{fips codes of states for lookup}
}
\value{
A dataframe of downloaded ACS data, in wide format.
}
\description{
Downloads and Formats ACS Tracts for the place geography
}
