% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAndFormatAcs.R
\name{downloadAndFormatAcs}
\alias{downloadAndFormatAcs}
\title{downloadAndFormatAcs}
\usage{
downloadAndFormatAcs(tables, geography = "state", year, survey = "acs5")
}
\arguments{
\item{tables}{array of strings, corresponding to the US Census tables to be downloaded}

\item{geography}{the geography of the request; can be "state" / "us" / "place".}

\item{year}{year of data to download}

\item{survey}{specification of ACS survey type (ex. "acs5" or "acs1")}
}
\value{
A dataframe of downloaded ACS data, in wide format.
}
\description{
Downloads and Formats ACS Data
}
