% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadCountryWideAcs.R
\name{downloadCountryWideAcs}
\alias{downloadCountryWideAcs}
\title{downloadCountryWideAcs}
\usage{
downloadCountryWideAcs(tables, geography = "state", year, survey = "acs5")
}
\arguments{
\item{tables}{array of strings, corresponding to the US Census tables to be downloaded}

\item{geography}{the geography of the request; can be geographies such as "state" / "us" / "place".}

\item{year}{year of data to download}

\item{survey}{specification of ACS survey type (ex. "acs5" or "acs1")}
}
\value{
A dataframe of downloaded ACS data, in wide format.
}
\description{
Downloads ACS Data for the full country, according to the specified geography. Please consult the list here to ensure that the geography you are requesting does not require a sub-geography, such as state, in order to request: https://walker-data.com/tidycensus/articles/basic-usage.html#geography-in-tidycensus
}
