% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSFOfCitiesPlaces.R
\name{downloadSFOfCitiesPlaces}
\alias{downloadSFOfCitiesPlaces}
\title{downloadSFOfCitiesPlaces}
\usage{
downloadSFOfCitiesPlaces(year, fips_codes_for_states)
}
\arguments{
\item{year}{year of data to download}

\item{fips_codes_for_states}{list of fips codes for states that we want to look up}
}
\value{
A dataframe of downloaded data for counties and places within the US
}
\description{
Downloads Spatial Footprint (SF) of Cities and Places in US, based on FIPS codes
}
