% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSFOfTracts.R
\name{downloadSFOfTracts}
\alias{downloadSFOfTracts}
\title{downloadSFOfTracts}
\usage{
downloadSFOfTracts(year, fips_codes_for_lookup)
}
\arguments{
\item{year}{year of data to download}

\item{fips_codes_for_lookup}{list of fips codes that we want to look up}
}
\value{
A dataframe of downloaded data for Census tracts within the US
}
\description{
Downloads Spatial Footprint (SF) of Tracts in US, based on FIPS codes
}
