% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSpatialFootprintOfCitiesPlaces.R
\name{downloadSpatialFootprintOfCitiesPlaces}
\alias{downloadSpatialFootprintOfCitiesPlaces}
\title{downloadSpatialFootprintOfCitiesPlaces}
\usage{
downloadSpatialFootprintOfCitiesPlaces(year)
}
\arguments{
\item{year}{year of data to download}
}
\value{
A dataframe of downloaded data for counties and places within the US
}
\description{
Downloads SF of Cities and Places in US, based on FIPS codes
}
