% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadStateWideAcs.R
\name{downloadStateWideAcs}
\alias{downloadStateWideAcs}
\title{downloadStateWideAcs}
\usage{
downloadStateWideAcs(
  tables,
  geography = "tract",
  year,
  survey = "acs5",
  state_codes_for_lookup
)
}
\arguments{
\item{tables}{array of strings, corresponding to the US Census tables to be downloaded}

\item{geography}{the geography of the request}

\item{year}{year of data to download}

\item{survey}{specification of ACS survey type (ex. "acs5" or "acs1")}

\item{state_codes_for_lookup}{list of fips codes of states for lookup}
}
\value{
A dataframe of downloaded ACS data, in wide format.
}
\description{
Downloads and Formats ACS Tracts by state
}
