% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPlaceAcsDataToScope.R
\name{filterPlaceAcsDataToScope}
\alias{filterPlaceAcsDataToScope}
\title{filterPlaceAcsDataToScope}
\usage{
filterPlaceAcsDataToScope(place_acs_data, geoids_in_scope)
}
\arguments{
\item{place_acs_data}{place acs data to be filtered}

\item{geoids_in_scope}{list of GEOIDs to filter the data}
}
\value{
A dataframe of filtered ACS data, in wide format.
}
\description{
filterPlaceAcsDataToScope
}
