% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPctAndNonPctData.R
\name{formatPctAndNonPctData}
\alias{formatPctAndNonPctData}
\title{formatPctAndNonPctData}
\usage{
formatPctAndNonPctData(df)
}
\arguments{
\item{df}{dataframe that is not yet formatted}
}
\value{
formatted df
}
\description{
Format percent and non-percent data consistently across all data
}
