% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountryWideAcsForMultipleYears.R
\name{getCountryWideAcsForMultipleYears}
\alias{getCountryWideAcsForMultipleYears}
\title{getCountryWideAcsForMultipleYears}
\usage{
getCountryWideAcsForMultipleYears(years, scope)
}
\arguments{
\item{years}{the years of data to retrieve}

\item{scope}{The scope of the geography to request--for example, can be "tracts" to get census tract data, or can be "national," which will request both "state" and "us" geographies from the cesus}
}
\value{
A dataframe of country-wide ACS data, in wide format, for both US and state-levels.
}
\description{
This function retrieves country-wide ACS data, using the file 'ada_parc_base_variables_mapped_to_source_variables.xlsx" to pull the base ACS variables that are specified there. This allows multiple years of data to be retrieved, if relevant.
}
