% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFIPSCodesForStateAndCounty.R
\name{getFIPSCodes}
\alias{getFIPSCodes}
\title{getFIPSCodes}
\usage{
getFIPSCodes()
}
\value{
A df including information on US FIPS codes, filtered for use in analysis
}
\description{
Downloads FIPS codes to use for analysis
}
