% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFIPSCodesForStateAndCounty.R
\name{getFIPSCodesForStateAndCounty}
\alias{getFIPSCodesForStateAndCounty}
\title{getFIPSCodesForStateAndCounty}
\usage{
getFIPSCodesForStateAndCounty()
}
\value{
A df of US FIPS codes for states and counties, filtered for use in analysis. Note that state codes are unique and county codes are unique WITHIN states.
}
\description{
Downloads FIPS codes for states and counties to use for analysis
}
