% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFIPSCodes.R
\name{getFIPSForSFs}
\alias{getFIPSForSFs}
\title{getFIPSForSFs}
\usage{
getFIPSForSFs()
}
\value{
A list of FIPS codes to be used to download SFs
}
\description{
Downloads FIPS codes to use for downloading SFs
}
