% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNationalHudForMultipleYears.R
\name{getNationalHudForMultipleYears}
\alias{getNationalHudForMultipleYears}
\title{getNationalHudForMultipleYears}
\usage{
getNationalHudForMultipleYears(years)
}
\arguments{
\item{years}{the years of data to retrieve}
}
\value{
A dataframe of country-wide ACS data, in wide format, for both US and state-levels.
}
\description{
Get HUD data for multiple years specified.
}
