% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectGeographiesSplitByState.R
\name{intersectGeographiesSplitByState}
\alias{intersectGeographiesSplitByState}
\title{intersectGeographiesSplitByState}
\usage{
intersectGeographiesSplitByState(
  geography_split_by_state_1,
  geography_split_by_state_2
)
}
\arguments{
\item{geography_split_by_state_1}{a dataframe (type SF) that is split by state}

\item{geography_split_by_state_2}{a dataframe (type SF) that is split by state}
}
\value{
a dataframe that combines
}
\description{
Intersect two SF dataframes that are split by states
}
