% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCensusAPIKey.R
\name{loadCensusAPIKey}
\alias{loadCensusAPIKey}
\title{loadCensusAPIKey}
\usage{
loadCensusAPIKey()
}
\value{
Census API Key
}
\description{
Load Census API Key
}
\details{
This allows for US Census Data to be downloaded. Here, we assume that someone already has a key present and we simply load it.
}
