% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRawExtractedDataFile.R
\name{readRawExtractedDataFile}
\alias{readRawExtractedDataFile}
\title{readRawExtractedDataFile}
\usage{
readRawExtractedDataFile(name)
}
\arguments{
\item{name}{name for the file}
}
\value{
the R object that was saved
}
\description{
Read raw extracted data file.
}
