% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTransformedDataFile.R
\name{readTransformedExtractedDataFile}
\alias{readTransformedExtractedDataFile}
\title{readTransformedExtractedDataFile}
\usage{
readTransformedExtractedDataFile(name)
}
\arguments{
\item{name}{name for the file}
}
\value{
the R object that was saved
}
\description{
Read processed/transformed extracted data file.
}
