% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveTransformedDataFile.R
\name{saveTransformedDataFile}
\alias{saveTransformedDataFile}
\title{saveTransformedDataFile}
\usage{
saveTransformedDataFile(transformed_data, name)
}
\arguments{
\item{transformed_data}{Transformed data as we manipulate and calculate new variables. Must be an R object.}

\item{name}{name for the file}
}
\description{
Save Transformed Data Files as Rds. Why Rds? W"You can save an R object like a data frame as either an RData file or an RDS file. RData files can store multiple R objects at once, but RDS files are the better choice because they foster reproducible code...For example, if you have three R objects, \code{a}, \code{b}, and \code{c}, you could save them all in the same RData file and then reload them in another R session...However, if you forget the names of your objects or give your file to someone else to use, it will be difficult to determine what was in the file—even after you (or they) load it. The user interface for RDS files is much more clear. You can save only one object per file, and whoever loads it can decide what they want to call their new data. As a bonus, you don’t have to worry about \code{load} overwriting any R objects that happened to have the same name as the objects you are loading:" For more information, read "D.4.1 Saving R Files" in  https://rstudio-education.github.io/hopr/dataio.html
}
