% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformNationalHudData.R
\name{transformNationalHudData}
\alias{transformNationalHudData}
\title{transformNationalHudData}
\usage{
transformNationalHudData(hud_data)
}
\arguments{
\item{hud_data}{raw hud data}
}
\value{
transformed hud data
}
\description{
transformNationalHudData
}
