% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/transformPlacesTractsCountiesWithFIPSAndTidyNames.R
\name{transformPlacesTractsCountiesWithFIPSAndTidyNames}
\alias{transformPlacesTractsCountiesWithFIPSAndTidyNames}
\title{transformPlacesTractsCountiesWithFIPSAndTidyNames}
\usage{
transformPlacesTractsCountiesWithFIPSAndTidyNames(
  places_tracts,
  county_state_fips
)
}
\arguments{
\item{places_tracts}{a map that has place, tract, and county GEOIDs plus place and tract names}

\item{county_state_fips}{dataframe containing places and their population}
}
\value{
a tidied place_tract_county_map
}
\description{
Combines a places_tracts map with FIPS codes and names, cleaning up the names of places for display. Currently, we don't use the county field; we include the county information in case we decide that we want county-wide analysis. Will save us effort down the line, if need be.
}
