% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSFOfCounties.R,
%   R/downloadSpatialFootprintOfCitiesPlaces.R
\name{downloadSpatialFootprintOfCitiesPlaces}
\alias{downloadSpatialFootprintOfCitiesPlaces}
\title{downloadSpatialFootprintOfCitiesPlaces}
\usage{
downloadSpatialFootprintOfCitiesPlaces(year, fips_codes_for_lookup)

downloadSpatialFootprintOfCitiesPlaces(year, fips_codes_for_lookup)
}
\arguments{
\item{year}{year of data to download}

\item{fips_codes_for_lookup}{list of fips codes that we want to look up}
}
\value{
A dataframe of downloaded data for counties and places within the US

A dataframe of downloaded data for counties and places within the US
}
\description{
Downloads SF of Cities and Places in US, based on FIPS codes

Downloads SF of Cities and Places in US, based on FIPS codes
}
