% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getADAPARCBaseToSourceVariableMap.R
\name{getADAPARCBaseToSourceVariableMap}
\alias{getADAPARCBaseToSourceVariableMap}
\title{getADAPARCBaseToSourceVariableMap}
\usage{
getADAPARCBaseToSourceVariableMap()
}
\value{
a dataframe mapping ADA-PARC base variables to their source variables, including the year and the source of the data (ex. ACS, HUD)
}
\description{
Read in a dataframe that maps ADA-PARC variables to their source
}
