% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFIPSCodesForStates.R
\name{getFIPSCodesForStates}
\alias{getFIPSCodesForStates}
\title{getFIPSCodesForStates}
\usage{
getFIPSCodesForStates()
}
\value{
A list of FIPS state codes
}
\description{
Downloads FIPS codes for states and counties to use for analysis
}
