% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectGeographiesSplitByState.R
\name{intersectGeographiesSplitByState}
\alias{intersectGeographiesSplitByState}
\title{intersectGeographiesSplitByState}
\usage{
intersectGeographiesSplitByState(places_sf, tracts_sf)
}
\arguments{
\item{places_sf}{a dataframe (type SF) that is split by state}

\item{tracts_sf}{a dataframe (type SF) that is split by state}
}
\value{
a dataframe that combines
}
\description{
Intersect two SF dataframes that are split by states. They will be intersected, left-right. The two dataframes must be of the same length.
}
